Divided the code into functions
